% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-harmo_process_harmonization.R
\name{as_data_proc_elem}
\alias{as_data_proc_elem}
\title{Validate and coerce any object as a Data Processing Elements}
\usage{
as_data_proc_elem(object)
}
\arguments{
\item{object}{A potential Data Processing Elements to be coerced.}
}
\value{
A tibble, identifying a Data Processing Elements.
}
\description{
Validates the input object as a valid Data Processing Elements and coerces
it with the appropriate Rmonize::class attribute. This function mainly
helps validate input within other functions of the package but could be used
to check if an object is valid for use in a function.
}
\details{
A Data Processing Elements contains the rules and metadata that will be used
to perform harmonization of input datasets in accordance with the DataSchema.
It must be a data frame or data frame extension (e.g. a tibble) and it must
contain certain columns which participate to the process, including the
\code{dataschema_variable}, \code{input_dataset},\code{input_variables},
\code{Mlstr_harmo::rule_category} and  \code{Mlstr_harmo::algorithm}. The mandatory
first processing element must be "id_creation" in
\code{Mlstr_harmo::rule_category} followed by the name of the column taken as
identifier of each dataset to initiate the process of harmonization.
}
\examples{
{

# You can use our demonstration files to run examples

as_data_proc_elem(DEMO_files_harmo$`data_processing_elements - final`)
}

}
