% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-harmo_process_harmonization.R
\name{as_dataschema_mlstr}
\alias{as_dataschema_mlstr}
\title{Validate and coerce any object as a DataSchema with Maelstrom restrictions}
\usage{
as_dataschema_mlstr(object)
}
\arguments{
\item{object}{A potential Maelstrom formatted DataSchema to be coerced.}
}
\value{
A list of tibble(s), each of them identifying the DataSchema.
}
\description{
Validates the input object as a valid DataSchema compliant with formats used
in Maelstrom Research ecosystem, including Opal, and returns it with the
appropriate Rmonize::class attribute. This function mainly helps validate
input within other functions of the package but could be used to check if an
object is valid for use in a function.
}
\details{
A DataSchema defines the harmonized variables to be generated, and also
represents metadata of an associated harmonized dossier. It must be a list
of data frame or data frame extension (e.g. a tibble) objects with elements
named "Variables" (required) and "Categories" (if any). The "Variables"
element must contain at least the \code{name} column, and the "Categories" element
must contain at least the \code{variable} and \code{name} columns to be usable in any
function. To be considered as a minimum workable DataSchema, in "Variables"
the \code{name} column must also have unique and non-null entries, and in
"Categories" the combination of \code{variable} and \code{name} columns must also be
unique.
}
\examples{
{

# You can use our demonstration files to run examples

as_dataschema_mlstr(Rmonize_DEMO$`dataschema - final`)

}

}
