% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-harmo_process_harmonization.R
\name{as_harmonized_dossier}
\alias{as_harmonized_dossier}
\title{Validate and coerce any object as an harmonized dossier}
\usage{
as_harmonized_dossier(
  object,
  dataschema = NULL,
  data_proc_elem = NULL,
  harmonized_col_id = NULL
)
}
\arguments{
\item{object}{A potential harmonized dossier to be coerced.}

\item{dataschema}{A list of tibble(s) representing metadata of an
associated harmonized dossier.}

\item{data_proc_elem}{A tibble, identifying the input
Data Processing Elements.}

\item{harmonized_col_id}{A character string identifying the name of the
column present in every dataset as identifier of the dataset.}
}
\value{
A list of tibble(s), each of them identifying the harmonized dataset.
}
\description{
Validates the input object as a valid harmonized dossier and coerces it with
the appropriate Rmonize::class attribute. This function mainly helps
validate input within other functions of the package but could be used to
check if an object is valid for use in a function.
}
\details{
A harmonized dossier must be a named list containing at least one data frame
or data frame extension (e.g. a tibble), each of them being
harmonized dataset(s). It is generally the product of applying harmonization
processing to a dossier object. The name of each tibble will be use as the
reference name of the dataset. A harmonized dossier has four attributes :
\code{Rmonize::class} which is "harmonized_dossier" ; \code{Rmonize::DataSchema}
(provided by user) ; \verb{Rmonize::Data Processing Elements} ;
\code{Rmonize::harmonized_col_id} (provided by user) which refers to the column
in each dataset which identifies unique combination observation/dataset.
This id column name is the same across the dataset(s), the DataSchema and
the Data Processing Elements (created by using 'id_creation') and is used to
initiate the process of harmonization.

A DataSchema defines the harmonized variables to be generated, and also
represents metadata of an associated harmonized dossier. It must be a list
of data frame or data frame extension (e.g. a tibble) objects with elements
named "Variables" (required) and "Categories" (if any). The "Variables"
element must contain at least the \code{name} column, and the "Categories" element
must contain at least the \code{variable} and \code{name} columns to be usable in any
function. To be considered as a minimum workable DataSchema, in "Variables"
the \code{name} column must also have unique and non-null entries, and in
"Categories" the combination of \code{variable} and \code{name} columns must also be
unique.

A Data Processing Elements contains the rules and metadata that will be used
to perform harmonization of input datasets in accordance with the DataSchema.
It must be a data frame or data frame extension (e.g. a tibble) and it must
contain certain columns which participate to the process, including the
\code{dataschema_variable}, \code{input_dataset},\code{input_variables},
\code{Mlstr_harmo::rule_category} and  \code{Mlstr_harmo::algorithm}. The mandatory
first processing element must be "id_creation" in
\code{Mlstr_harmo::rule_category} followed by the name of the column taken as
identifier of each dataset to initiate the process of harmonization.
}
\examples{
{

as_harmonized_dossier(DEMO_files_harmo$harmonized_dossier)
  
}

}
