% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-import-from-madshapR.R
\docType{import}
\name{data_dict_extract}
\alias{data_dict_extract}
\title{Objects exported from other packages}
\keyword{imported}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{madshapR}{\code{\link[madshapR]{data_dict_extract}}}
}}

