% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-harmonized_data_evaluate.R
\name{data_proc_elem_evaluate}
\alias{data_proc_elem_evaluate}
\title{Generate a quality assessment report of a Data Processing Elements}
\usage{
data_proc_elem_evaluate(data_proc_elem, taxonomy = NULL)
}
\arguments{
\item{data_proc_elem}{A tibble, identifying the input
Data Processing Elements.}

\item{taxonomy}{A tibble identifying the scheme used for variables
classification.}
}
\value{
A list of tibbles of report for one Data Processing Elements.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Function that assesses the content and structure of a
Data Processing Elements object and reports possible issues to facilitate
assessment of input data. The report can be used to help assess
data structure, presence of fields, coherence across elements, and taxonomy
or data dictionary formats. This report is compatible with Excel and can be
exported as an Excel spreadsheet.
}
\details{
A Data Processing Elements contains the rules and metadata that will be used
to perform harmonization of input datasets in accordance with the DataSchema.
It must be a data frame or data frame extension (e.g. a tibble) and it must
contain certain columns which participate to the process, including the
\code{dataschema_variable}, \code{input_dataset},\code{input_variables},
\code{Mlstr_harmo::rule_category} and  \code{Mlstr_harmo::algorithm}. The mandatory
first processing element must be "id_creation" in
\code{Mlstr_harmo::rule_category} followed by the name of the column taken as
identifier of each dataset to initiate the process of harmonization.

A taxonomy is classification scheme that can be defined for variable
attributes. If defined, a taxonomy must be a data frame like object. It must
be compatible with (and is generally extracted from) an Opal environment. To
work with certain functions, a valid taxonomy must contain at least the
columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy
may follow Maelstrom research taxonomy, and its content can be evaluated
accordingly, such as naming convention restriction, tagging elements,
or scales, which are specific to Maelstrom Research. In this particular
case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale',
'vocabulary_scale' and 'term_scale' to work with some specific functions.
}
\examples{
{

# use Rmonize_DEMO provided by the package

data_proc_elem <- Rmonize_DEMO$`data_processing_elements - final`   
data_proc_elem_evaluate(data_proc_elem)

}

}
