% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-import-from-madshapR.R
\docType{import}
\name{dataset_summarize}
\alias{dataset_summarize}
\title{Objects exported from other packages}
\keyword{imported}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{madshapR}{\code{\link[madshapR]{dataset_summarize}}}
}}

