% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\docType{data}
\name{DEMO_files_harmo}
\alias{DEMO_files_harmo}
\title{Built-in material allowing the user to test the package with demo data}
\format{
\subsection{\code{list}}{

A list with 15 elements used for testing the package (data frames and lists):

\describe{
\item{DEMO_data_processing_elements - final}{Data Processing Elements}
\item{DEMO_data_processing_elements - with error}{Data Processing Elements
containing errors}
\item{DEMO_data_processing_elements - work in progress}{Data processing
element in construction}
\item{DEMO_dataschema}{DataSchema used, combined with data processing
elements and input material}
\item{pooled_harmonized_dataset}{pooled harmonized dataset for testing
purpose}
\item{harmonized_dossier}{harmonized dossier for testing purpose}
\item{summary_var_harmo}{harmonized variables summary for testing purpose}
\item{dd_MELBOURNE_1_format_maelstrom}{Data dictionary (1) of Melbourne
dataset}
\item{dd_MELBOURNE_2_format_maelstrom}{Data dictionary (2) of Melbourne
dataset}
\item{dd_PARIS_format_maelstrom}{Data dictionary of Paris dataset}
\item{dd_TOKYO_format_maelstrom}{Data dictionary of Tokyo dataset}
\item{dataset_MELBOURNE_1}{Dataset of Melbourne (1)}
\item{dataset_MELBOURNE_2}{Dataset of Melbourne (2)}
\item{dataset_PARIS}{Dataset of Paris}
\item{dataset_TOKYO}{Dataset of Tokyo}
...
}
}
}
\usage{
DEMO_files_harmo
}
\description{
Built-in tibbles and lists allowing the user to test the package with demo
material for harmonization.
}
\examples{
{

 print(DEMO_files_harmo$`dataschema - final`)

}
}
\keyword{datasets}
