% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-harmonized_data_evaluate.R
\name{harmonized_dossier_evaluate}
\alias{harmonized_dossier_evaluate}
\title{Generate a report of a harmonized dossier (list of harmonized datasets)}
\usage{
harmonized_dossier_evaluate(harmonized_dossier, taxonomy = NULL)
}
\arguments{
\item{harmonized_dossier}{List of tibble(s), each of them being
harmonized dataset.}

\item{taxonomy}{A tibble identifying the scheme used for variables
classification.}
}
\value{
A list of report(s), each of them being tibble(s) ('Overview and summary)
for each harmonized dataset.
}
\description{
Assesses the content and structure of a harmonized dossier object (list of
harmonized datasets) and reports possible issues in the datasets and
data dictionaries  to facilitate assessment of input data. The report can be
used to help assess data structure, presence of fields, coherence across
elements, and taxonomy or data dictionary formats. This report is compatible
with Excel and can be exported as an Excel spreadsheet.
}
\details{
A harmonized dossier must be a named list containing at least one data frame
or data frame extension (e.g. a tibble), each of them being
harmonized dataset(s). It is generally the product of applying harmonization
processing to a dossier object. The name of each tibble will be use as the
reference name of the dataset. A harmonized dossier has four attributes :
\code{Rmonize::class} which is "harmonized_dossier" ; \code{Rmonize::DataSchema}
(provided by user) ; \verb{Rmonize::Data Processing Elements} ;
\code{Rmonize::harmonized_col_id} (provided by user) which refers to the column
in each dataset which identifies unique combination observation/dataset.
This id column name is the same across the dataset(s), the DataSchema and
the Data Processing Elements (created by using 'id_creation') and is used to
initiate the process of harmonization.

A taxonomy is classification scheme that can be defined for variable
attributes. If defined, a taxonomy must be a data frame like object. It must
be compatible with (and is generally extracted from) an Opal environment. To
work with certain functions, a valid taxonomy must contain at least the
columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy
may follow Maelstrom research taxonomy, and its content can be evaluated
accordingly, such as naming convention restriction, tagging elements,
or scales, which are specific to Maelstrom Research. In this particular
case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale',
'vocabulary_scale' and 'term_scale' to work with some specific functions.
}
\examples{
{

harmonized_dossier_evaluate(DEMO_files_harmo$harmonized_dossier)

}

}
