% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-import-from-madshapR.R
\docType{import}
\name{is_data_dict_mlstr}
\alias{is_data_dict_mlstr}
\title{Objects exported from other packages}
\keyword{imported}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{madshapR}{\code{\link[madshapR]{is_data_dict_mlstr}}}
}}

