% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-harmo_process_harmonization.R
\name{is_dataschema_mlstr}
\alias{is_dataschema_mlstr}
\title{Test if an object is a valid Maelstrom dataSchema}
\usage{
is_dataschema_mlstr(object)
}
\arguments{
\item{object}{A potential Maelstrom formatted DataSchema to be evaluated.}
}
\value{
A logical.
}
\description{
Tests if the input object is a valid dataSchema compliant with formats
used in Maelstrom Research ecosystem, including Opal. This function mainly
helps validate input within other functions of the package but could be used
to check if an object is valid for use in a function.
}
\details{
A DataSchema defines the harmonized variables to be generated, and also
represents metadata of an associated harmonized dossier. It must be a list
of data frame or data frame extension (e.g. a tibble) objects with elements
named "Variables" (required) and "Categories" (if any). The "Variables"
element must contain at least the \code{name} column, and the "Categories" element
must contain at least the \code{variable} and \code{name} columns to be usable in any
function. To be considered as a minimum workable DataSchema, in "Variables"
the \code{name} column must also have unique and non-null entries, and in
"Categories" the combination of \code{variable} and \code{name} columns must also be
unique.
}
\examples{
{

# use Rmonize_DEMO provided by the package

dataschema <- Rmonize_DEMO$`dataschema - final`
is_dataschema_mlstr(dataschema)
is_dataschema_mlstr(iris)

}

}
\seealso{
For a better assessment, please use \code{\link[=dataschema_evaluate]{dataschema_evaluate()}}.
}
