% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-harmo_process_harmonization.R
\name{pooled_harmonized_dataset_create}
\alias{pooled_harmonized_dataset_create}
\title{Generate the pooled dataset from harmonized dataset(s) in a dossier}
\usage{
pooled_harmonized_dataset_create(
  harmonized_dossier,
  harmonized_col_dataset =
    attributes(harmonized_dossier)$`Rmonize::harmonized_col_dataset`,
  harmonized_col_id = attributes(harmonized_dossier)$`Rmonize::harmonized_col_id`,
  add_col_dataset = FALSE,
  dataschema = attributes(harmonized_dossier)$`Rmonize::DataSchema`,
  data_proc_elem = attributes(harmonized_dossier)$`Rmonize::Data Processing Elements`,
  dataschema_apply = FALSE
)
}
\arguments{
\item{harmonized_dossier}{List of tibble(s), each of them being
harmonized dataset.}

\item{harmonized_col_dataset}{A character string identifying the name of the
column present in every dataset as identifier of the dataset name.}

\item{harmonized_col_id}{A character string identifying the name of the
column identifier of the dataset and will be the concatenation of
id column value and dataset name. NULL by default.}

\item{add_col_dataset}{whether to add an extra column to each
harmonized dataset. The resulting tibble will have an additional column and
its data dictionary will be updated accordingly adding categories for this
variable. FALSE by default.}

\item{dataschema}{A list of tibble(s) representing metadata of an
associated harmonized dossier.}

\item{data_proc_elem}{A tibble, identifying the input
Data Processing Elements.}

\item{dataschema_apply}{whether to apply the datashema to each
harmonized dataset. The resulting tibble will have for each column its
associated meta data as attributes. The factors will be preserved.
FALSE by default.}
}
\value{
A tibble, which is the pooled harmonized dataset from a harmonized dossier.
}
\description{
Generates the pooled harmonized dataset from harmonized dataset(s) in a
dossier. The pooled dataset has two columns which can be declared by
the user (harmonized_col_dataset and harmonized_col_id).
The first column refers to the name of each dataset which is the name of each
tibble in the dossier. The second column refers to the column id in each
harmonized dataset and  which identifies unique combination (concatenated)
observation/dataset. These two columns are added to ensure every information
is safe during the process. The pooled harmonized dataset comes with its
data dictionary which is the harmonized dossier DataSchema, to which the
two extra columns are added.
}
\details{
A harmonized dossier must be a named list containing at least one data frame
or data frame extension (e.g. a tibble), each of them being
harmonized dataset(s). It is generally the product of applying harmonization
processing to a dossier object. The name of each tibble will be use as the
reference name of the dataset. A harmonized dossier has four attributes :
\code{Rmonize::class} which is "harmonized_dossier" ; \code{Rmonize::DataSchema}
(provided by user) ; \verb{Rmonize::Data Processing Elements} ;
\code{Rmonize::harmonized_col_id} (provided by user) which refers to the column
in each dataset which identifies unique combination observation/dataset.
This id column name is the same across the dataset(s), the DataSchema and
the Data Processing Elements (created by using 'id_creation') and is used to
initiate the process of harmonization.
}
\examples{
{

harmonized_dossier <- Rmonize_DEMO$harmonized_dossier

pooled_harmonized_dataset_create(
 harmonized_dossier,harmonized_col_id = 'adm_unique_id')
  
}

}
