% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-import-from-madshapR.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_data_dict}
\alias{is_data_dict}
\alias{as_data_dict_mlstr}
\alias{is_data_dict_mlstr}
\alias{as_dataset}
\alias{is_dataset}
\alias{as_dossier}
\alias{is_dossier}
\alias{data_extract}
\alias{data_dict_extract}
\alias{data_dict_apply}
\alias{dataset_zap_data_dict}
\alias{dossier_create}
\alias{dataset_evaluate}
\alias{data_dict_evaluate}
\alias{dossier_evaluate}
\alias{dataset_summarize}
\alias{dossier_summarize}
\alias{dataset_visualize}
\alias{valueType_adjust}
\alias{as_taxonomy}
\alias{bookdown_template}
\alias{bookdown_render}
\alias{bookdown_open}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{madshapR}{\code{\link[madshapR]{as_data_dict}}, \code{\link[madshapR]{as_data_dict_mlstr}}, \code{\link[madshapR]{as_dataset}}, \code{\link[madshapR]{as_dossier}}, \code{\link[madshapR]{as_taxonomy}}, \code{\link[madshapR:reexports]{bookdown_open}}, \code{\link[madshapR:reexports]{bookdown_render}}, \code{\link[madshapR:reexports]{bookdown_template}}, \code{\link[madshapR]{data_dict_apply}}, \code{\link[madshapR]{data_dict_evaluate}}, \code{\link[madshapR]{data_dict_extract}}, \code{\link[madshapR]{data_extract}}, \code{\link[madshapR]{dataset_evaluate}}, \code{\link[madshapR]{dataset_summarize}}, \code{\link[madshapR]{dataset_visualize}}, \code{\link[madshapR]{dataset_zap_data_dict}}, \code{\link[madshapR]{dossier_create}}, \code{\link[madshapR]{dossier_evaluate}}, \code{\link[madshapR]{dossier_summarize}}, \code{\link[madshapR]{is_data_dict}}, \code{\link[madshapR]{is_data_dict_mlstr}}, \code{\link[madshapR]{is_dataset}}, \code{\link[madshapR]{is_dossier}}, \code{\link[madshapR]{valueType_adjust}}}
}}

