% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\docType{data}
\name{Rmonize_DEMO}
\alias{Rmonize_DEMO}
\title{Built-in material allowing the user to test the package with demo data}
\format{
\subsection{\code{list}}{

A list with 15 elements used for testing the package (data frames and lists):

\describe{
\item{data_processing_elements - final}{Data Processing Elements}
\item{data_processing_elements - with error}{Data Processing Elements
containing errors}
\item{data_processing_elements - work in progress}{Data processing
element in construction}
\item{dataschema}{DataSchema used, combined with data processing
elements and input material}
\item{harmonized_dossier}{harmonized dossier for testing purpose}
\item{harmonized_dossier_summary}{harmonized variables summary for testing
purpose}
\item{pooled_harmonized_dataset}{pooled harmonized dataset for testing
purpose}
\item{data_dict_MELBOURNE}{Data dictionary of Melbourne
dataset}
\item{data_dict_PARIS}{Data dictionary of Paris dataset}
\item{data_dict_TOKYO}{Data dictionary of Tokyo dataset}
\item{dataset_MELBOURNE}{Dataset of Melbourne}
\item{dataset_PARIS}{Dataset of Paris}
\item{dataset_TOKYO}{Dataset of Tokyo}
...
}
}
}
\usage{
Rmonize_DEMO
}
\description{
Built-in tibbles and lists allowing the user to test the package with demo
material for harmonization.
}
\examples{
{

 print(Rmonize_DEMO$`dataschema - final`)

}
}
\keyword{datasets}
