% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\docType{data}
\name{Rmonize_examples}
\alias{Rmonize_examples}
\title{Objects to provide illustrative use case}
\format{
\subsection{\code{list}}{

A list with 12 elements (data frames and lists) providing example objects
for testing the package:

\describe{
\item{dataset_study1}{Dataset example for a study named "study1", labelled with its data dictionary}
\item{dataset_study2}{Dataset example for a study named "study2", labelled with its data dictionary}
\item{dataset_study3}{Dataset example for a study named "study3", labelled with its data dictionary}
\item{dataset_study4}{Dataset example for a study named "study4", labelled with its data dictionary}
\item{dataset_study5}{Dataset example for a study named "study5", labelled with its data dictionary}
\item{Data Processing Elements}{Example Data Processing Elements}
\item{DataSchema}{Example DataSchema}
\item{harmo_dataset_study1}{Harmonized dataset example , labelled with its harmonized data dictionary}
\item{harmo_dataset_study2}{Harmonized dataset example , labelled with its harmonized data dictionary}
\item{harmo_dataset_study3}{Harmonized dataset example , labelled with its harmonized data dictionary}
\item{harmo_dataset_study4}{Harmonized dataset example , labelled with its harmonized data dictionary}
\item{harmo_dataset_study5}{Harmonized dataset example , labelled with its harmonized data dictionary}
\item{harmo_dataset_pooled}{Harmonized pooled dataset example, labelled with its harmonized data dictionary}
\item{summary - harmonized datasets}{Example of harmonized datasets summary}
...
}
}
}
\usage{
Rmonize_examples
}
\description{
Illustrative input datasets, input data dictionaries, DataSchema,
Data Processing Elements, and other objects to provide use case of Rmonize.
}
\examples{
{

library(dplyr)

glimpse(Rmonize_examples$DataSchema)

}
}
\keyword{datasets}
