% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\name{Rmonize_help}
\alias{Rmonize_help}
\title{Call to online documentation}
\usage{
Rmonize_help()
}
\value{
Nothing to be returned. The function opens a web page.
}
\description{
Direct call to the online documentation for the package, which includes a
description of the latest version of the package, vignettes, user guides,
and a reference list of functions and help pages.
}
\examples{
{

Rmonize_help()

}

}
