% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\name{Rmonize_templates}
\alias{Rmonize_templates}
\title{Call the help center to the download template page}
\usage{
Rmonize_templates()
}
\value{
Nothing to be returned. The function opens a web page.
}
\description{
This function is a direct call to the documentation in the repository hosting
the package, more specifically to the download template section.
The user can also access the description of the latest version of the
package, the vignettes, and the list of functions.
}
\examples{
{

download_templates()

}

}
