% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-harmo_process_harmonization.R
\name{show_harmo_error}
\alias{show_harmo_error}
\title{Generate a summary of the annotated Data Processing Elements in the console}
\usage{
show_harmo_error(harmonized_dossier)
}
\arguments{
\item{harmonized_dossier}{List of tibble(s), each of them being
harmonized dataset.}
}
\value{
No object returned, the function sends messages in the console, showing
errors in the harmonization process.
}
\description{
Reads annotated Data Processing Elements to list processes, any errors, and
an overview of each harmonization rule. This summary report can be used to
asses harmonization and provide insight on the data structure, fields to
investigate, and coherence across elements.
}
\details{
A harmonized dossier must be a named list containing at least one data frame
or data frame extension (e.g. a tibble), each of them being
harmonized dataset(s). It is generally the product of applying harmonization
processing to a dossier object. The name of each tibble will be use as the
reference name of the dataset. A harmonized dossier has four attributes :
\code{Rmonize::class} which is "harmonized_dossier" ; \code{Rmonize::DataSchema}
(provided by user) ; \verb{Rmonize::Data Processing Elements} ;
\code{Rmonize::harmonized_col_id} (provided by user) which refers to the column
in each dataset which identifies unique combination observation/dataset.
This id column name is the same across the dataset(s), the DataSchema and
the Data Processing Elements (created by using 'id_creation') and is used to
initiate the process of harmonization.
}
\examples{
{

harmonized_dossier <- Rmonize_DEMO$harmonized_dossier
show_harmo_error(harmonized_dossier)
}

}
