% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-harmonized_data_visualize.R
\name{harmonized_dossier_visualize}
\alias{harmonized_dossier_visualize}
\title{Generate a web-based visual report for a harmonized dossier}
\usage{
harmonized_dossier_visualize(
  harmonized_dossier,
  bookdown_path,
  group_by = attributes(harmonized_dossier)$`Rmonize::harmonized_col_dataset`,
  harmonized_dossier_summary = NULL,
  dataschema = attributes(harmonized_dossier)$`Rmonize::DataSchema`,
  data_proc_elem = attributes(harmonized_dossier)$`Rmonize::Data Processing Element`,
  valueType_guess = FALSE,
  taxonomy = NULL
)
}
\arguments{
\item{harmonized_dossier}{A list containing the harmonized dataset(s).}

\item{bookdown_path}{A character string identifying the folder path where
the bookdown report files will be saved.}

\item{group_by}{A character string identifying the column in each
harmonized dataset to use as a grouping variable. Visual elements will be
grouped by this column.}

\item{harmonized_dossier_summary}{A list which identifies an existing
summary produced by \code{\link[=harmonized_dossier_summarize]{harmonized_dossier_summarize()}} of the harmonized
variables.
Using this parameter can save time in generating the visual report.}

\item{dataschema}{A DataSchema object.}

\item{data_proc_elem}{A Data Processing Elements object.}

\item{valueType_guess}{Whether the output should be generated based on more
precise valueType inferred from the data. FALSE by default
(will use the valueType declared).}

\item{taxonomy}{An optional data frame identifying a variable
classification schema.}
}
\value{
A folder containing files for the bookdown site. To open the bookdown site
in a browser, open 'docs/index.html', or use \code{\link[=bookdown_open]{bookdown_open()}} with the
folder path.
}
\description{
Generates a visual report of a harmonized dossier in an HTML bookdown
document, with summary figures and statistics for each harmonized variable.
The report outputs can be grouped by a categorical variable.
}
\details{
A harmonized dossier is a named list containing one or more data frames,
which are harmonized datasets. A harmonized dossier is generally the
product of applying processing to a dossier object The name of each
harmonized dataset (data frame) is taken from the reference input dataset.
A harmonized dossier also contains the DataSchema and
Data Processing Elements used in processing as attributes.

A DataSchema is the list of core variables to generate across datasets and
related metadata. A DataSchema object is a list of data frames with elements
named 'Variables' (required) and 'Categories' (if any). The 'Variables'
element must contain at least the \code{name} column, and the 'Categories'
element must contain at least the \code{variable} and \code{name} columns to be usable
in any function. In 'Variables' the \code{name} column must also have unique
entries, and in 'Categories' the combination of \code{variable} and \code{name} columns
must also be unique.

The Data Processing Elements specifies the algorithms used to process input
variables into harmonized variables in the DataSchema format. It is also
contains metadata used to generate documentation of the processing.
A Data Processing Elements object is a data frame with specific columns
used in data processing: \code{dataschema_variable}, \code{input_dataset},
\code{input_variables}, \code{Mlstr_harmo::rule_category} and \code{Mlstr_harmo::algorithm}.
To initiate processing, the first entry must be the creation of a harmonized
primary identifier variable (e.g., participant unique ID).

The valueType is a declared property of a variable that is required in
certain functions to determine handling of the variables. Specifically,
valueType refers to the
\href{https://opaldoc.obiba.org/en/variables-data.html#value-types}{OBiBa data type of a variable}.
The valueType is specified in a data dictionary in a column 'valueType' and
can be associated with variables as attributes. Acceptable valueTypes
include 'text', 'integer', 'decimal', 'boolean', datetime', 'date'. The full
list of OBiBa valueType possibilities and their correspondence with R data
types are available using \link{valueType_list}.

A taxonomy is a classification schema that can be defined for variable
attributes. A taxonomy is usually extracted from an
\href{https://www.obiba.org/pages/products/opal/}{Opal environment}, and a
taxonomy object is a data frame that must contain at least the columns
\code{taxonomy}, \code{vocabulary}, and \code{terms}. Additional details about Opal
taxonomies are
\href{https://opaldoc.obiba.org/en/latest/web-user-guide/administration/taxonomies.html}{available online}.
}
\examples{
{

# Use Rmonize_DEMO to run examples.

library(fs)

harmonized_dossier <- Rmonize_DEMO$harmonized_dossier
harmonized_dossier_summary <- Rmonize_DEMO$harmonized_dossier_summary

if(dir_exists(tempdir())) dir_delete(tempdir())
bookdown_path <- tempdir()

harmonized_dossier_visualize(
  harmonized_dossier,
  bookdown_path = bookdown_path,
  harmonized_dossier_summary = harmonized_dossier_summary)

# To open the file in browser, open 'bookdown_path/docs/index.html'.
# Or use bookdown_open(bookdown_path) function

}

}
\seealso{
\code{\link[=dataset_visualize]{dataset_visualize()}}
\code{\link[=bookdown_open]{bookdown_open()}}
}
