% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-harmonized_data_visualise.R
\name{harmonized_dossier_visualize}
\alias{harmonized_dossier_visualize}
\title{Generate a web-based bookdown visual report of a harmonized dossier}
\usage{
harmonized_dossier_visualize(
  harmonized_dossier,
  bookdown_path,
  group_by = attributes(harmonized_dossier)$`Rmonize::harmonized_col_dataset`,
  harmonized_dossier_summary = NULL,
  dataschema = attributes(harmonized_dossier)$`Rmonize::DataSchema`,
  data_proc_elem = attributes(harmonized_dossier)$`Rmonize::Data Processing Element`,
  valueType_guess = FALSE,
  taxonomy = NULL
)
}
\arguments{
\item{harmonized_dossier}{List of tibble(s), each of them being
harmonized dataset.}

\item{bookdown_path}{A character string identifying the folder path where the
bookdown report will be saved.}

\item{group_by}{A character string of one column in the dataset that can be
taken as a grouping column. The visual element will be grouped and displayed
by this column.}

\item{harmonized_dossier_summary}{A list which is the summary of the
harmonized dossier.}

\item{dataschema}{A list of tibble(s) representing metadata of an
associated harmonized dossier.}

\item{data_proc_elem}{A tibble, identifying the input
Data Processing Elements.}

\item{valueType_guess}{Whether the output should include a more accurate
valueType that could be applied to the dataset. FALSE by default.
The visual element will be grouped and displayed by this column.}

\item{taxonomy}{A tibble identifying the scheme used for variables
classification.}
}
\value{
A bookdown folder containing files in the specified output folder. To
open the file in browser, open 'docs/index.html'.
Or use \code{\link[=bookdown_open]{bookdown_open()}}
}
\description{
Generates a visual report for a dataset in an HTML bookdown document. The
report provides figures and descriptive statistics for each variable to
facilitate the assessment of input data. Statistics and figures are generated
according to variable data type. The report can be used to help assess
data structure, coherence across elements, and taxonomy or
data dictionary formats. The summaries and figures provide additional
information about variable distributions and descriptive statistics.
The charts and tables are produced based on their data type. The variable can
be grouped using \code{group_by} parameter, which is a (categorical) column in the
dataset. The user may need to use \code{\link[=as.factor]{as.factor()}} in this context. To fasten
the process (and allow recycling object in a workflow) the user can feed the
function with a \code{.summary_var}, which is the output of the function
\code{\link[=dataset_summarize]{dataset_summarize()}} of the column(s) \code{col} and  \code{group_by}. The summary
must have the same parameters to operate.
}
\details{
A harmonized dossier must be a named list containing at least one data frame
or data frame extension (e.g. a tibble), each of them being
harmonized dataset(s). It is generally the product of applying harmonization
processing to a dossier object. The name of each tibble will be use as the
reference name of the dataset. A harmonized dossier has four attributes :
\code{Rmonize::class} which is "harmonized_dossier" ; \code{Rmonize::DataSchema}
(provided by user) ; \verb{Rmonize::Data Processing Elements} ;
\code{Rmonize::harmonized_col_id} (provided by user) which refers to the column
in each dataset which identifies unique combination observation/dataset.
This id column name is the same across the dataset(s), the DataSchema and
the Data Processing Elements (created by using 'id_creation') and is used to
initiate the process of harmonization.

A taxonomy is classification scheme that can be defined for variable
attributes. If defined, a taxonomy must be a data frame like object. It must
be compatible with (and is generally extracted from) an Opal environment. To
work with certain functions, a valid taxonomy must contain at least the
columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy
may follow Maelstrom research taxonomy, and its content can be evaluated
accordingly, such as naming convention restriction, tagging elements,
or scales, which are specific to Maelstrom Research. In this particular
case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale',
'vocabulary_scale' and 'term_scale' to work with some specific functions.

The valueType is a property of a variable and is required in certain
functions to determine the handling of the variables. The valueType refers
to the OBiBa-internal type of a variable. It is specified in a data
dictionary in a column \code{valueType} and can be associated with variables as
attributes. Acceptable valueTypes include 'text', 'integer', 'decimal',
'boolean', datetime', 'date'). The full list of OBiBa valueType
possibilities and their correspondence with R data types are available using
\link[madshapR:valueType_list]{madshapR::valueType_list}.
}
\examples{
{

# You can use our demonstration files to run examples

library(fs)

harmonized_dossier <- Rmonize_DEMO$harmonized_dossier
harmonized_dossier_summary <- Rmonize_DEMO$harmonized_dossier_summary

if(dir_exists(tempdir())) dir_delete(tempdir())
bookdown_path <- tempdir()

harmonized_dossier_visualize(
  harmonized_dossier,
  bookdown_path = bookdown_path,
  harmonized_dossier_summary = harmonized_dossier_summary)

# To open the file in browser, open 'bookdown_path/docs/index.html'.
# Or use bookdown_open(bookdown_path) function

}

}
\seealso{
\code{\link[=pooled_harmonized_dataset_create]{pooled_harmonized_dataset_create()}}
\code{\link[=dataset_summarize]{dataset_summarize()}}
\code{\link[=bookdown_open]{bookdown_open()}}
}
