% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-import-from-madshapR.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{valueType_adjust}
\alias{as_taxonomy}
\alias{is_taxonomy}
\alias{bookdown_open}
\alias{bookdown_render}
\alias{is_data_dict}
\alias{is_data_dict_mlstr}
\alias{is_dataset}
\alias{is_dossier}
\alias{data_extract}
\alias{dataset_zap_data_dict}
\alias{dossier_create}
\alias{as_valueType}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{madshapR}{\code{\link[madshapR]{as_taxonomy}}, \code{\link[madshapR]{as_valueType}}, \code{\link[madshapR]{bookdown_open}}, \code{\link[madshapR]{bookdown_render}}, \code{\link[madshapR]{data_extract}}, \code{\link[madshapR]{dataset_zap_data_dict}}, \code{\link[madshapR]{dossier_create}}, \code{\link[madshapR]{is_data_dict}}, \code{\link[madshapR]{is_data_dict_mlstr}}, \code{\link[madshapR]{is_dataset}}, \code{\link[madshapR]{is_dossier}}, \code{\link[madshapR]{is_taxonomy}}, \code{\link[madshapR]{valueType_adjust}}}
}}

