% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-import-from-madshapR.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_data_dict_mlstr}
\alias{as_taxonomy}
\alias{as_valueType}
\alias{as_category}
\alias{is_category}
\alias{data_extract}
\alias{dataset_zap_data_dict}
\alias{is_data_dict}
\alias{is_data_dict_mlstr}
\alias{is_dataset}
\alias{is_dossier}
\alias{is_taxonomy}
\alias{valueType_adjust}
\alias{valueType_self_adjust}
\alias{col_id}
\alias{data_dict_filter}
\alias{data_dict_trim_labels}
\alias{data_dict_update}
\alias{has_categories}
\alias{first_label_get}
\alias{data_dict_match_dataset}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{madshapR}{\code{\link[madshapR]{as_category}}, \code{\link[madshapR]{as_data_dict_mlstr}}, \code{\link[madshapR]{as_taxonomy}}, \code{\link[madshapR]{as_valueType}}, \code{\link[madshapR]{col_id}}, \code{\link[madshapR]{data_dict_filter}}, \code{\link[madshapR]{data_dict_match_dataset}}, \code{\link[madshapR]{data_dict_trim_labels}}, \code{\link[madshapR]{data_dict_update}}, \code{\link[madshapR]{data_extract}}, \code{\link[madshapR]{dataset_zap_data_dict}}, \code{\link[madshapR]{first_label_get}}, \code{\link[madshapR]{has_categories}}, \code{\link[madshapR]{is_category}}, \code{\link[madshapR]{is_data_dict}}, \code{\link[madshapR]{is_data_dict_mlstr}}, \code{\link[madshapR]{is_dataset}}, \code{\link[madshapR]{is_dossier}}, \code{\link[madshapR]{is_taxonomy}}, \code{\link[madshapR]{valueType_adjust}}, \code{\link[madshapR]{valueType_self_adjust}}}
}}

