% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmonize-package.R
\docType{package}
\name{Rmonize-package}
\alias{Rmonize}
\alias{Rmonize-package}
\title{Rmonize: Tools for Data Harmonization}
\description{
Integrated tools to support rigorous and well documented data harmonization based on Maelstrom Research guidelines. The package includes functions to assess and prepare input elements, apply specified processing rules to generate harmonized datasets, validate data processing and identify processing errors, and document and summarize harmonized outputs. The harmonization process is defined and structured by two key user-generated documents: the DataSchema (specifying the list of harmonized variables to generate across datasets) and the Data Processing Elements (specifying the input elements and processing algorithms to generate harmonized variables in DataSchema formats). The package was developed to address key challenges of retrospective data harmonization in epidemiology (as described in Fortier I and al. (2017) \doi{10.1093/ije/dyw075}) but can be used for any data harmonization initiative.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maelstrom-research/Rmonize/}
  \item Report bugs at \url{https://github.com/maelstrom-research/Rmonize/issues}
}

}
\author{
\strong{Maintainer}: Guillaume Fabre \email{guijoseph.fabre@gmail.com} (\href{https://orcid.org/0000-0002-0124-9970}{ORCID})

Authors:
\itemize{
  \item Maelstrom Research \email{info@maelstrom-research.org} [funder, copyright holder]
}

}
\keyword{internal}
