% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\docType{data}
\name{Rmonize_examples}
\alias{Rmonize_examples}
\title{Objects to provide illustrative use case}
\format{
\subsection{\code{list}}{

A list with 12 elements (data frames and lists) providing example objects
for testing the package:

\describe{
\item{dataset_study1}{Example Dataset for a study named "study1", labelled with its data dictionary}
\item{dataset_study2}{Example Dataset for a study named "study2", labelled with its data dictionary}
\item{dataset_study3}{Example Dataset for a study named "study3", labelled with its data dictionary}
\item{dataset_study4}{Example Dataset for a study named "study4", labelled with its data dictionary}
\item{dataset_study5}{Example Dataset for a study named "study5", labelled with its data dictionary}
\item{T3_PAE_use_case_2024_08_21}{Example Data Processing Elements}
\item{T2_PAE_use_case_2024_08_21}{Example DataSchema}
...
}
}
}
\usage{
Rmonize_examples
}
\description{
Illustrative input datasets, input data dictionaries, DataSchema,
Data Processing Elements, and other objects to provide use case of Rmonize.
}
\examples{
{

library(dplyr)

glimpse(Rmonize_examples$DataSchema)

}
}
\keyword{datasets}
