% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\docType{data}
\name{Rmonize_examples}
\alias{Rmonize_examples}
\title{Example objects to provide an illustrative use case}
\format{
\subsection{\code{list}}{

A list with elements (data frames and lists) providing example objects
for using the package:
\describe{
\item{original_dataset_study1}{Example original dataset from .sav file for "study1"}
\item{original_dataset_study2}{Example original dataset from .sav file for "study2"}
\item{original_dataset_study3}{Example original dataset from .csv file for "study3"}
\item{original_dataset_study4}{Example original dataset from .xlsx file for "study4"}
\item{original_dataset_study5}{Example original dataset from .xlsx file for "study5"}
\item{original_data_dictionary_study4}{Example original data dictionary from .xlsx file, for "study4"}
\item{original_data_dictionary_study5}{Example original data dictionary from .xlsx file, for "study5"}
\item{input_dataset_study1}{Example input dataset ready for processing, for "study1"}
\item{input_dataset_study2}{Example input dataset ready for processing, for "study2"}
\item{input_dataset_study3}{Example input dataset ready for processing, for "study3"}
\item{input_dataset_study4}{Example input dataset ready for processing, for "study4"}
\item{input_dataset_study5}{Example input dataset ready for processing, for "study5"}
\item{DataSchema}{Example DataSchema}
\item{Data_Processing_Elements_no_errors}{Example Data Processing Elements containing no errors}
\item{Data_Processing_Elements_with_errors}{Example Data Processing Elements containing errors}
\item{harmonized_dossier}{Example harmonized dossier}
\item{pooled_harmonized_dataset}{Example pooled harmonized dataset}
\item{summary_report_harmonized_dossier}{Example summary report of harmonized dossier}
}
}
}
\usage{
Rmonize_examples
}
\description{
Example input datasets, input data dictionaries, DataSchema,
Data Processing Elements, harmonized output, and summary report.
}
\examples{
{

library(dplyr)
glimpse(Rmonize_examples$`DataSchema`)


}
}
\keyword{datasets}
