% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\name{Rmonize_templates}
\alias{Rmonize_templates}
\title{Call to online documentation to download templates}
\usage{
Rmonize_templates()
}
\value{
Nothing to be returned. The function opens a web page.
}
\description{
Direct call to online documentation to download templates.
}
\examples{
{

Rmonize_templates()

}

}
