% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\name{Rmonize_website}
\alias{Rmonize_website}
\title{Call to package website}
\usage{
Rmonize_website()
}
\value{
Nothing to be returned. The function opens a web page.
}
\description{
Direct call to the package website, which includes an overview of the
Rmonize process, vignettes and user guides, a reference list of functions
and help pages, and package updates.
}
\examples{
{

Rmonize_website()

}

}
